+++ tests/tree_map/test_tree_map_iterator.c
@@ -184,7 +184,7 @@ check_static_begin(tree_map_test_forward_iterator) {
     check(j, 0);
     int const num_nodes = 33;
     int const prime = 37;
-    size_t shuffled_index = prime % num_nodes;
+    size_t shuffled_index = (size_t)prime % (size_t)num_nodes;
     for (int i = 0; i < num_nodes; ++i) {
         (void)swap_entry(
             &s,
@@ -197,10 +197,10 @@ check_static_begin(tree_map_test_forward_iterator) {
             &allocator
         );
         check(validate(&s), true);
-        shuffled_index = (shuffled_index + prime) % num_nodes;
+        shuffled_index = (shuffled_index + (size_t)prime) % (size_t)num_nodes;
     }
     int val_keys_inorder[33];
-    check(inorder_fill(val_keys_inorder, num_nodes, &s), CHECK_PASS);
+    check(inorder_fill(val_keys_inorder, (size_t)num_nodes, &s), CHECK_PASS);
     j = 0;
     for (struct Val *e = begin(&s); e && j < num_nodes;
          e = next(&s, &e->elem), ++j) {
